# FRONTEND/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BACKEND",
    "version": "1.0"
  },
  "paths": {
    "/accountlist": {
      "get": {
        "tags": [
          "Account"
        ],
        "operationId": "AccountList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/account/add": {
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "AddAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/account/Edit": {
      "put": {
        "tags": [
          "Account"
        ],
        "operationId": "EditAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/account/del/{accountId}": {
      "delete": {
        "tags": [
          "Account"
        ],
        "operationId": "DeleteAccount",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/ddl": {
      "get": {
        "tags": [
          "Dropdown"
        ],
        "operationId": "DropdownUserAndMaster",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllDdl"
                }
              }
            }
          }
        }
      }
    },
    "/ddlreport": {
      "get": {
        "tags": [
          "Dropdown"
        ],
        "operationId": "DropdownUserAndMasterRepport",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllDdl"
                }
              }
            }
          }
        }
      }
    },
    "/masterlist": {
      "get": {
        "tags": [
          "Master"
        ],
        "operationId": "MasterList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MasterTransaction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/master/add": {
      "post": {
        "tags": [
          "Master"
        ],
        "operationId": "AddMaster",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasterDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/master/Edit": {
      "put": {
        "tags": [
          "Master"
        ],
        "operationId": "EditMaster",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasterDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/master/del/{masterId}": {
      "delete": {
        "tags": [
          "Master"
        ],
        "operationId": "DeleteMaster",
        "parameters": [
          {
            "name": "masterId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/reportpoints": {
      "get": {
        "tags": [
          "Report"
        ],
        "operationId": "ReportPoints",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/reportall": {
      "get": {
        "tags": [
          "Report"
        ],
        "operationId": "ReportAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transactionlist": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "operationId": "TransactionList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transaction/add": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "operationId": "AddTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/transaction/Edit": {
      "put": {
        "tags": [
          "Transaction"
        ],
        "operationId": "EditTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/transaction/del/{transId}": {
      "delete": {
        "tags": [
          "Transaction"
        ],
        "operationId": "DeleteTransaction",
        "parameters": [
          {
            "name": "transId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AllDdl": {
        "type": "object",
        "properties": {
          "userList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListDdl"
            },
            "nullable": true
          },
          "masterList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListDdl"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ListDdl": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MasterDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MasterTransaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TransDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "accountId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "accountName": {
            "type": "string",
            "nullable": true
          },
          "transactionDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "masterTransactionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "debitCreditStatus": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "points": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "accountId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "account": {
            "$ref": "#/components/schemas/Account"
          },
          "transactionDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "masterTransactionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "masterTransaction": {
            "$ref": "#/components/schemas/MasterTransaction"
          },
          "debitCreditStatus": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "point": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}